Refactor 'ContainsAnyOrdinal' to do a quick probe first
Move 'QuickSubstringProbe' from 'StringBuilderExtensions' to 'StringExtensions', make it an extension method and rename it to 'QuickSustringProbeOrdinal'
