+++ MiKo.Analyzer.Shared/Rules/Documentation/OverallDocumentationAnalyzer.cs
@@ -81,7 +81,7 @@ namespace MiKoSolutions.Analyzers.Rules.Documentation
 
         protected IReadOnlyList<Diagnostic> AnalyzeComment(
                                                        DocumentationCommentTriviaSyntax comment,
-                                                       IEnumerable<string> termsForText,
+                                                       in ReadOnlySpan<string> termsForText,
                                                        in ReadOnlySpan<string> termsForAllLocations,
                                                        string replacement = "",
                                                        in StringComparison comparison = StringComparison.OrdinalIgnoreCase,
